refactor: 분리가 가능한 DOM 조작 로직 분리
